fix exmath decorator
