+++ src/watex/tools/funcutils.py
@@ -22,10 +22,6 @@ import pandas as pd
 import matplotlib.pyplot as plt
 
 from .._watexlog import watexlog
-from ..decorators import ( 
-    deprecated, 
-    catmapflow2,
-    )
 
 from ..typing import ( 
     Tuple,
@@ -1770,144 +1766,7 @@ def get_boundaries(df):
     
     return _autoOption, shape, type_, ves_loc , posMinMax,  df.iloc[:, :4]  
     
-@deprecated ('Function should be removed for the next release.')
-def get_type (
-        erp_array,
-        posMinMax,
-        pk, pos_array,
-        dl
-              ): 
-    """
-    Find anomaly type from app. resistivity values and positions locations 
-    
-    :param erp_array: App.resistivty values of all `erp` lines 
-    :type erp_array: array_like 
-    
-    :param posMinMax: Selected anomaly positions from startpoint and endpoint 
-    :type posMinMax: list or tuple or nd.array(1,2)
-    
-    :param pk: Position of selected anomaly in meters 
-    :type pk: float or int 
-    
-    :param pos_array: Stations locations or measurements positions 
-    :type pos_array: array_like 
-    
-    :param dl: 
-        
-        Distance between two receiver electrodes measurement. The same 
-        as dipole length in meters. 
-    
-    :returns: 
-        - ``EC`` for Extensive conductive. 
-        - ``NC`` for narrow conductive. 
-        - ``CP`` for conductive plane 
-        - ``CB2P`` for contact between two planes. 
-        
-    :Example: 
-        
-        >>> from watex.tools.funcutils import get_type 
-        >>> x = [60, 61, 62, 63, 68, 65, 80,  90, 100, 80, 100, 80]
-        >>> pos= np.arange(0, len(x)*10, 10)
-        >>> ano_type= get_type(erp_array= np.array(x),
-        ...            posMinMax=(10,90), pk=50, pos_array=pos, dl=10)
-        >>> ano_type
-        ...CB2P
-
-    """
-    # Get position index 
-    anom_type ='CP'
-    index_pos = int(np.where(pos_array ==pk)[0])
-    # if erp_array [:index_pos +1].mean() < np.median(erp_array) or\
-    #     erp_array[index_pos:].mean() < np.median(erp_array) : 
-    #         anom_type ='CB2P'
-    if erp_array [:index_pos+1].mean() < np.median(erp_array) and \
-        erp_array[index_pos:].mean() < np.median(erp_array) : 
-            anom_type ='CB2P'
-            
-    elif erp_array [:index_pos +1].mean() >= np.median(erp_array) and \
-        erp_array[index_pos:].mean() >= np.median(erp_array) : 
-                
-        if  dl <= (max(posMinMax)- min(posMinMax)) <= 5* dl: 
-            anom_type = 'NC'
-
-        elif (max(posMinMax)- min(posMinMax))> 5 *dl: 
-            anom_type = 'EC'
-
-    return anom_type   
-        
-    
-@catmapflow2(cat_classes=['FR0', 'FR1', 'FR2', 'FR3'])#, 'FR4'] )
-def categorize_flow(
-        target_array: T ,
-        flow_values: List [float],
-        **kwargs
-    ) -> Tuple[ List[float], T, List[str]]: 
-    """ 
-    Categorize `flow` into different classes. If the optional
-    `flow_classes`  argument is given, it should be erased the
-    `cat_classes` argument of decororator `deco.catmapflow`.
-    
-    :param target_array: Flow array to be categorized 
-    
-    :param flow_values: 
-        
-        The way to be categorized. Distribute the flow values 
-        of numerical values considered as pseudo_classes like:: 
-    
-            flow_values= [0.0, [0.0, 3.0], [3.0, 6.0], [6.0, 10.0], 10.0] (1)
-            
-        if ``flow_values`` is given as follow:: 
-            
-            flow_values =[0. , 3., 6., 10.] (2)
-        
-        It should convert the type (2) to (1).
-        
-    :param flow_classes: 
-        Values of categorized flow rates 
-        
-    :returns: 
-        
-        - ``new_flow_values``: Iterable object as type (2) 
-        - ``target_array``: Raw flow iterable object to be categorized
-        - ``flowClasses``: If given , see ``flow_classes`` param. 
-            
-    """
-    flowClasses =  kwargs.pop('classes', None)
-
-    new_flow_values = []
-    inside_inter_flag= False
-    
-    if isinstance(flow_values, (tuple, np.ndarray)): 
-        flow_values =list(flow_values)
-    # Loop and find 
-    for jj, _iter in enumerate(flow_values) : 
-        if isinstance(_iter, (list, tuple, np.ndarray)): 
-            inside_inter_flag = True 
-            flow_values[jj]= list(_iter)
  
-    if inside_inter_flag: 
-        new_flow_values =flow_values 
-    
-    if inside_inter_flag is False: 
-        flow_values= sorted(flow_values)
-        # if 0. in flow_values : 
-        #     new_flow_values.append(0.) 
-        for ss, val in enumerate(flow_values) : 
-            if ss ==0 : 
-                #append always the first values. 
-                 new_flow_values.append(val) 
-            # if val !=0. : 
-            else:
-                if val ==flow_values[-1]: 
-                    new_flow_values.append([flow_values[ss-1], val])
-                    new_flow_values.append(val)
-                else: 
-                   new_flow_values.append([flow_values[ss-1], val])
- 
-    return new_flow_values, target_array, flowClasses
-    
-    
-   
 def reshape(arr , axis = None) :
     """ Detect the array shape and reshape it accordingly, back to the given axis.
